Connect Enemy refactoring
