+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/Boss.java
@@ -1,22 +1,17 @@
 package com.cursedarchie.platformer.actors.enemies;
 
 import com.badlogic.gdx.math.Vector2;
-import com.cursedarchie.platformer.actors.Enemy;
+import com.cursedarchie.platformer.actors.NewEnemy;
 
-public class Boss extends Enemy {
+public class Boss extends NewEnemy {
 
     public Boss(Vector2 pos) {
-        super(pos);
-        this.setSize(4f);
-        this.setMaxHealth(10f);
+        super(pos, 4f, 10f);
         this.setMaxAcceleration(10f);
         this.setDamage(1f);
         this.setAttackDuration(2f);
         this.setViewDistance(5f);
         this.setViewAngle(90f);
-        this.setHealth(this.getMaxHealth());
-        this.getBounds().setWidth(this.getSize());
-        this.getBounds().setHeight(this.getSize());
     }
 
 }
