+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/logic/states/DeadState.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.math.Vector2;
 import com.cursedarchie.platformer.actors.NewEnemy;
 import com.cursedarchie.platformer.actors.enemies.logic.EnemyState;
 
-public class DyingState implements EnemyState {
+public class DeadState implements EnemyState {
     @Override
     public void enter(NewEnemy enemy) {
         Gdx.app.log("Enemy", "Started Dying");
@@ -17,11 +17,10 @@ public class DyingState implements EnemyState {
         enemy.setDamage(0);
         enemy.setPosition(new Vector2(99, 99));
         enemy.setAlive(false);
-        enemy.getStateMachine().changeState(null);
     }
 
     @Override
     public void exit (NewEnemy enemy) {
-        Gdx.app.log("Enemy", "Ended Dying");
+        Gdx.app.log("Enemy", "Finished Dying");
     }
 }
