+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/logic/states/PatrolState.java
@@ -1,12 +1,13 @@
 package com.cursedarchie.platformer.actors.enemies.logic.states;
 
+import com.badlogic.gdx.Gdx;
 import com.cursedarchie.platformer.actors.NewEnemy;
 import com.cursedarchie.platformer.actors.enemies.logic.EnemyState;
 
 public class PatrolState implements EnemyState {
     @Override
     public void enter(NewEnemy enemy) {
-
+        Gdx.app.log("Enemy", "Started Patroling");
     }
 
     @Override
@@ -16,6 +17,6 @@ public class PatrolState implements EnemyState {
 
     @Override
     public void exit (NewEnemy enemy) {
-
+        Gdx.app.log("Enemy", "Finished Patroling");
     }
 }
